+++ src/components/routes/TRouteCreate.php
@@ -4,6 +4,7 @@ namespace extas\components\routes;
 use extas\components\exceptions\AlreadyExist;
 use extas\components\extensions\TExtendable;
 use extas\components\Plugins;
+use extas\interfaces\stages\IStageApiBeforeCreate;
 use extas\interfaces\stages\IStageApiValidateInputData;
 use Psr\Http\Message\ResponseInterface;
 use ReflectionClass;
@@ -28,9 +29,12 @@ trait TRouteCreate
             return $this->response;
         }
 
-        $class = $this->{$this->repoName}()->getItemClass();
-        $item = new $class($data);
         try {
+            $this->enrichData($data);
+
+            $class = $this->{$this->repoName}()->getItemClass();
+            $item = new $class($data);
+        
             $item = $this->{$this->repoName}()->create($item);
         } catch (\Exception $e) {
             $this->setResponseData($data, $e->getMessage());
@@ -40,6 +44,17 @@ trait TRouteCreate
         $this->setResponseData($item->__toArray());
 
         return $this->response;
+    }
+
+    protected function enrichData(array &$data): void
+    {
+        foreach(Plugins::byStage(IStageApiBeforeCreate::NAME) as $plugin) {
+            $plugin($data);
+        }
+
+        foreach(Plugins::byStage(IStageApiBeforeCreate::NAME . '.' . $this->repoName) as $plugin) {
+            $plugin($data);
+        }
     }
 
     protected function isValidData(array $data): bool