+++ tests/resources/TestDeleteDispatcher.php
@@ -17,6 +17,10 @@ class TestDeleteDispatcher extends JsonDispatcher
 
     protected function getWhere(): array
     {
+        if ($this->getRequestParameter('fail', '') == 'yes') {
+            throw new \Exception('Fail');
+        }
+        
         return [IRoute::FIELD__NAME => '/'];
     }
 }